requirements.txt
asgiref==3.2.10
certifi==2020.6.20
chardet==3.0.4
Django==3.0.8
django-mysql==3.7.1
django-utils==0.0.2
djangorestframework==3.11.0
djangotoolbox==1.8.0
idna==2.10
mysqlclient==2.0.1
psycopg2-binary==2.8.5
pytz==2020.1
requests==2.24.0
sqlparse==0.3.1
urllib3==1.25.9
psycopg2>=2.8